+++ src/api/base.ts
@@ -61,7 +61,6 @@ export class BaseAPI {
   }
 
   patch(id: string | number, data: any, apiPath?: string) {
-    console.log('PATCH');
     return this.http.patch(this.getPath(apiPath) + id + '/', data);
   }
