+++ src/api/container-distribution.ts
@@ -2,10 +2,6 @@ import { PulpAPI } from './pulp';
 
 export class API extends PulpAPI {
   apiPath = 'distributions/container/container/';
-
-  constructor() {
-    super();
-  }
 }
 
 export const ContainerDistributionAPI = new API();
