+++ src/api/execution-environment-namespace.ts
@@ -2,10 +2,6 @@ import { HubAPI } from './hub';
 
 class API extends HubAPI {
   apiPath = this.getUIPath('execution-environments/namespaces/');
-
-  constructor() {
-    super();
-  }
 }
 
 export const ExecutionEnvironmentNamespaceAPI = new API();
