+++ src/containers/execution-environment-detail/execution_environment_detail.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { withRouter } from 'react-router-dom';
-import { Main, EmptyStateNoData, MarkdownEditor } from '../../components';
+import { EmptyStateNoData, MarkdownEditor } from '../../components';
 import {
-  Main,
   MarkdownEditor,
   EmptyStateNoData,
 } from '../../components';
@@ -19,16 +18,12 @@ import { withContainerRepo, IDetailSharedProps } from './base';
 import {
   ExecutionEnvironmentAPI,
   GroupObjectPermissionType,
-  ExecutionEnvironmentNamespaceAPI,
 } from '../../api';
 import './execution-environment-detail.scss';
-import { RepositoryForm } from '../../components';
-import { ContainerDistributionAPI } from 'src/api/container-distribution';
 
 interface IState {
   loading: boolean;
   readme: string;
-  editing: boolean;
   markdownEditing: boolean;
   redirect: string;
   distribution_id: string;
@@ -47,7 +42,6 @@ class ExecutionEnvironmentDetail extends React.Component<
     this.state = {
       loading: true,
       readme: '',
-      editing: false,
       markdownEditing: false,
       redirect: null,
       distribution_id: '',
@@ -62,30 +56,7 @@ class ExecutionEnvironmentDetail extends React.Component<
   }
 
   render() {
-    return <Main><RepositoryForm
-        name={this.props.match.params['container']}
-        selectedGroups={this.state.groups}
-        description={this.state.description}
-        onSave={(description, selectedGroups) => {
-            let promises = [];
-            promises.push(
-                ContainerDistributionAPI.patch(this.state.distribution_id, {
-                    description: description,
-                }),
-            );
-            promises.push(
-                ExecutionEnvironmentNamespaceAPI.update(
-                    this.state.namespace.name,
-                    { groups: selectedGroups },
-                ),
-            );
-
-            Promise.all(promises).then(() =>
-                this.setState({ editing: false, description: description }),
-            );
-        }}
-        onCancel={() => this.setState({ editing: false })}
-    />{this.renderDetail()}</Main>;
+    return this.renderDetail();
   }
 
   renderDetail() {
